% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHFunctions.R
\name{addCalculatedVariables}
\alias{addCalculatedVariables}
\title{Add calculated variables}
\usage{
addCalculatedVariables(master_data)
}
\arguments{
\item{master_data}{JH compatiable dataframe}
}
\value{
dataframe with calculated variable added
}
\description{
Add calculated variables
}
\examples{
#TODO
}
