% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHFunctions.R
\name{addMissingCountryWideData}
\alias{addMissingCountryWideData}
\title{Add missing country wide data}
\usage{
addMissingCountryWideData(JH_Data)
}
\arguments{
\item{JH_Data}{dataframe containing expected John Hopkins data}
}
\value{
dataframe containing expected John Hopkins data with country wide
data added for countries that have only regional data - currently Canada,
Australia and China
}
\description{
Add missing country wide data
}
\examples{
#TODO
}
