% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHFunctions.R
\name{addPopulationData}
\alias{addPopulationData}
\title{Add Population Data}
\usage{
addPopulationData(master_data)
}
\arguments{
\item{master_data}{data frame containing JH compatible countries}
}
\value{
data frame with population information for each contry
}
\description{
Add Population Data
}
\examples{
#TODO
}
