% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_population.R
\docType{data}
\name{countries_population}
\alias{countries_population}
\title{World Population by Countries}
\format{
A data frame with 243 rows and 2 variables:
\describe{
  \item{Country_Region}{Country name, a character}
  \item{Population}{population of the country, a numeric}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_countries_and_dependencies_by_population}
}
\usage{
countries_population
}
\description{
A dataset containing the population of world countries with names compatible
with John Hopkins Data
}
\keyword{datasets}
