% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHFunctions.R
\name{getDataFromWeb}
\alias{getDataFromWeb}
\title{Gets data from web}
\usage{
getDataFromWeb(destfolder = tempdir())
}
\arguments{
\item{destfolder}{destination folder for downloaded files.}
}
\value{
master_data data frame containing deaths, confirmed, recovered and
vaccine data from John Hopkins
}
\description{
Gets data from web
}
\examples{
#TODO
}
