% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHFunctions.R
\name{getLatestAvailableDate}
\alias{getLatestAvailableDate}
\title{get latest available date in JH data}
\usage{
getLatestAvailableDate(JH_Data, CountryName)
}
\arguments{
\item{JH_Data}{JHData format dataframe}

\item{CountryName}{character valid country name}
}
\value{
latest available date
}
\description{
get latest available date in JH data
}
\examples{
#TODO
}
