% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHFunctions.R
\name{getLatestData}
\alias{getLatestData}
\title{get latest available JH Data over a days span}
\usage{
getLatestData(JH_Data, CountryList, spanDays = FALSE)
}
\arguments{
\item{JH_Data}{JH formatted data source (raw or augmented)}

\item{CountryList}{character list of countries to be extracted}

\item{spanDays}{number of days to look back start from latest available date in JH_Data.
If unspecified, defaults to a span of one day (latest available date)}
}
\value{
dataframe with the latest available data as specified by parameters
}
\description{
get latest available JH Data over a days span
}
\examples{
#TODO
}
