% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHFunctions.R
\name{getRawDataFromWeb}
\alias{getRawDataFromWeb}
\title{getRawDataFromWeb}
\usage{
getRawDataFromWeb(destfolder = tempdir())
}
\arguments{
\item{destfolder}{destination folder for downloaded files.}
}
\value{
master_data data frame containing deaths, confirmed, recovered and
vaccine data from John Hopkins, without change to regional data
}
\description{
getRawDataFromWeb
}
