% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHFunctions.R
\name{JHGetdata_CountryLevel}
\alias{JHGetdata_CountryLevel}
\title{Get Data at Country Level}
\usage{
JHGetdata_CountryLevel(JH_Data, CountryList, VarName)
}
\arguments{
\item{JH_Data}{data frame containing JH Data}

\item{CountryList}{character list of countries}

\item{VarName}{variable in JH_Data to be extracted}
}
\value{
ggplot plot
}
\description{
Get Data at Country Level
}
\examples{
#TODO
}
