% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHFunctions.R
\name{JHGetplot_CountryLevel}
\alias{JHGetplot_CountryLevel}
\title{Get ggplot plot of Single Variables for a list of Country}
\usage{
JHGetplot_CountryLevel(
  JH_Data,
  CountryList,
  VarName,
  add_label = FALSE,
  adjust_label = FALSE,
  mindiffval = 5
)
}
\arguments{
\item{JH_Data}{data frame containing JH Data}

\item{CountryList}{character list of countries}

\item{VarName}{character variable name in JH_Data to be extracted}

\item{add_label}{boolean on whether to add a label to the curve}

\item{adjust_label}{boolean on whether to adjust the label}

\item{mindiffval}{minimum value difference assured in case of label adjustmment}
}
\value{
ggplot of the variable with colour by country
}
\description{
Get ggplot plot of Single Variables for a list of Country
}
\examples{
#TODO
}
