% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHFunctions.R
\name{JHGetplot_CountryLevel_MultipleVar}
\alias{JHGetplot_CountryLevel_MultipleVar}
\title{Get facet_wrap ggplot plot of Multiple Variables for a list of Country}
\usage{
JHGetplot_CountryLevel_MultipleVar(JH_Data, CountryList, VarNames)
}
\arguments{
\item{JH_Data}{data frame containing JH Data}

\item{CountryList}{character list of countries}

\item{VarNames}{character list of variables in JH_Data to be extracted}
}
\value{
ggplot plot
}
\description{
Get facet_wrap ggplot plot of Multiple Variables for a list of Country
}
\examples{
#TODO
}
