% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHFunctions.R
\name{JHGetProgress}
\alias{JHGetProgress}
\title{Get latest variation}
\usage{
JHGetProgress(JH_Aug_Data, CountryList, VarName)
}
\arguments{
\item{JH_Aug_Data}{JH formatted data source (raw or augmented)}

\item{CountryList}{character list of countries to be extracted}

\item{VarName}{character JHData compatible variable name}
}
\value{
dataframe
}
\description{
Get latest variation
}
\examples{
#TODO
}
